Updated MR2 flags. This is in prep for supporting monster ring use.
Converted mintrinsics and mextrinsics both to longs to make room for more properties.

I'm following EvilHack's example here but taking it one small commit at a time.

+++ include/monst.h
@@ -113,8 +113,8 @@ struct monst {
     uchar m_ap_type;      /* what mappearance is describing, m_ap_types */
 
     schar mtame;                /* level of tameness, implies peaceful */
-    unsigned short mintrinsics; /* low 8 correspond to mresists */
-    unsigned short mextrinsics; /* low 8 correspond to mresists */
+    unsigned long mintrinsics; /* low 8 correspond to mresists */
+    unsigned long mextrinsics; /* low 8 correspond to mresists */
     unsigned long seen_resistance; /* M_SEEN_x; saw you resist an effect */
     int mspec_used;             /* monster's special ability attack timeout */
